bridge-ib-portal/pyproject.toml
[project]
name = "bridge-ib-portal"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "MetaTrader5>=5.0.45",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "cryptography>=42.0.0",
]
